requirements.txt
sgf>=0.5
pyyaml>=3.12
Pillow>=4.0.0
xvfbwrapper==0.2.9
Jinja2==2.9.6
